+++ src/Symfony/Component/Security/Core/Role/RoleHierarchy.php
@@ -18,6 +18,13 @@ namespace Symfony\Component\Security\Core\Role;
  */
 class RoleHierarchy implements RoleHierarchyInterface
 {
+    /**
+     * Map role placeholders with their regex pattern.
+     *
+     * @var array<string,string>
+     */
+    protected array $rolePlaceholdersPatterns;
+
     /** @var array<string, list<string>> */
     protected array $map;
 
@@ -32,24 +39,14 @@ class RoleHierarchy implements RoleHierarchyInterface
 
     public function getReachableRoleNames(array $roles): array
     {
-        $reachableRoles = $roles;
-
-        foreach ($roles as $role) {
-            if (!isset($this->map[$role])) {
-                continue;
-            }
-
-            foreach ($this->map[$role] as $r) {
-                $reachableRoles[] = $r;
-            }
-        }
-
-        return array_values(array_unique($reachableRoles));
+        return array_values(array_unique($this->resolveReachableRoleNames($roles)));
     }
 
     protected function buildRoleMap(): void
     {
         $this->map = [];
+        $this->rolePlaceholdersPatterns = [];
+
         foreach ($this->hierarchy as $main => $roles) {
             $this->map[$main] = $roles;
             $visited = [];
@@ -71,6 +68,66 @@ class RoleHierarchy implements RoleHierarchyInterface
             }
 
             $this->map[$main] = array_unique($this->map[$main]);
+
+            if (str_contains($main, '*') && null !== ($pattern = $this->getPlaceholderPattern($main))) {
+                $this->rolePlaceholdersPatterns[$main] = $pattern;
+            }
         }
+    }
+
+    private function resolveReachableRoleNames(array $roles, array &$visitedPlaceholders = []): array
+    {
+        $reachableRoles = $roles;
+
+        foreach ($roles as $role) {
+            if (!isset($this->map[$role])) {
+                continue;
+            }
+
+            foreach ($this->map[$role] as $r) {
+                $reachableRoles[] = $r;
+            }
+        }
+
+        $placeholderRoles = array_diff($this->getMatchingPlaceholders($reachableRoles), $visitedPlaceholders);
+        if (!empty($placeholderRoles)) {
+            array_push($visitedPlaceholders, ...$placeholderRoles);
+            $resolvedPlaceholderRoles = $this->resolveReachableRoleNames($placeholderRoles, $visitedPlaceholders);
+            foreach (array_diff($resolvedPlaceholderRoles, $placeholderRoles) as $r) {
+                $reachableRoles[] = $r;
+            }
+        }
+
+        return $reachableRoles;
+    }
+
+    protected function getMatchingPlaceholders(array $roles): array
+    {
+        $resolved = [];
+
+        foreach ($this->rolePlaceholdersPatterns as $placeholder => $pattern) {
+            if (!\in_array($placeholder, $resolved) && \count(preg_grep($pattern, $roles) ?: [])) {
+                $resolved[] = $placeholder;
+            }
+        }
+
+        return $resolved;
+    }
+
+    /**
+     * Build the regex pattern for the given role:
+     *   - Replace valid wildcards with a non-wildcard matching pattern.
+     *   - Escape reserved regex characters.
+     *
+     * A valid wildcard is a * prefixed with _ and immediately followed by _ or EOL.
+     *
+     * @return string|null The regex pattern, or null if there is no valid wildcard in the role
+     */
+    private function getPlaceholderPattern(string $role): ?string
+    {
+        /** @var int $count */
+        $placeholderPattern = preg_replace(pattern: '/(?<=_)\\\\\*(?=_|$)/', replacement: '.*', subject: preg_quote($role), count: $count);
+
+        return ($count > 0) ? sprintf('/^%s$/', $placeholderPattern) : null;
     }
 }